requirements.txt
absl-py==2.4.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
ale-py==0.11.2
annotated-types==0.7.0
anyio==4.11.0
array-api-compat==1.13.0
attrs==25.4.0
Box2D==2.3.10
certifi==2025.10.5
charset-normalizer==3.4.4
chex==0.1.91
click==8.3.0
cloudpickle==3.1.2
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.3.4
cycler==0.12.1
dataclasses-json==0.6.7
datasets==4.4.1
decorator==5.2.1
dill==0.4.0
distro==1.9.0
etils==1.13.0
faiss-cpu==1.13.0
Farama-Notifications==0.0.4
filelock==3.20.0
flax==0.12.4
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.10.0
glfw==2.10.0
grpcio==1.78.0
gymnasium==1.2.3
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.2
humanize==4.15.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
importlib_resources==6.5.2
jax==0.9.0.1
jaxlib==0.9.0.1
jaxtyping==0.3.7
Jinja2==3.1.6
jiter==0.11.1
kiwisolver==1.4.9
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.2
mujoco==3.5.0
multidict==6.7.0
multiprocess==0.70.18
mypy_extensions==1.1.0
neo4j==6.0.3
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.3.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
openai==2.7.1
opencv-python==4.13.0.92
opt_einsum==3.4.0
optax==0.2.6
orbax-checkpoint==0.11.32
orbax-export==0.0.8
packaging==25.0
pandas==2.3.3
pillow==11.3.0
proglog==0.1.12
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
pyarrow==22.0.0
pydantic==2.12.4
pydantic_core==2.41.5
pygame==2.6.1
Pygments==2.19.2
PyOpenGL==3.1.10
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.3.2
scipy==1.17.0
setuptools==69.5.1
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
stable_baselines3==2.7.1
swig==4.4.1
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorstore==0.1.81
toolz==1.1.0
torch==2.10.0
tqdm==4.67.1
treescope==0.1.10
triton==3.6.0
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wadler_lindig==0.1.7
Werkzeug==3.1.5
wheel==0.46.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
